fix stdint

+++ template/stdint.h
@@ -3,11 +3,11 @@
 typedef signed char int8_t;
 typedef signed short int int16_t;
 typedef signed int int32_t;
-typedef signed long long int int64_t;
+typedef signed long int int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long int uint64_t;
+typedef unsigned long int uint64_t;
 typedef uint8_t bool;
 
 #endif /* __STDINT_H */
